test : test 도메인별 분리
